// apps/web/generated/api/models/projectDetailsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationResponse } from './applicationResponse';
import type { DeliveryResponse } from './deliveryResponse';
import type { ProjectResponse } from './projectResponse';

export interface ProjectDetailsResponse {
  project: ProjectResponse;
  applications: ApplicationResponse[];
  deliveries: DeliveryResponse[];
}